Use BSON instead CBOR to marshal data
BSON is the only binary JSON-like data format that is well supported in almost all languages and platforms, thanks to the corporate (MongoDB) behind it.

I also decided to cancel the planned feature of generating language bindings to Rust structs. It is a lot of work and is of little use for targets of dynamically typed languages.

From now on, the generated target code will accept BSON values as all function parameters and return type. User will have to handle runtime errors like typo of field names.